StorageKeyRotationwithKeyVault/requirements.txt
azure-functions
azure-identity
azure-keyvault
azure-keyvault-secrets